Added Triangulation Import/Export functions
chiSurfaceMeshExportPolyFile. Wavefront.obj functions may contain edge information that can be used in the Delaunay Triangle mesher to create Constrained Triangulations. In order to run triangle1.6 in stand-alone mode, it needs to be supplied with a .poly file. A function has been added for this.

chiSurfaceMeshImportFromTriangleFiles. Triangle also produces its output as .node and .ele files for which an import utility has been created.