Modify SetSource for additive behaviors.
This commit allows for ClassicRichardson to initialize source iterations
with non-zero values. This is done by storing the current value of
`q_moments_local` before entering the source iteration loop. This is
primarily used for the KEigenvalue class where each power iteration
the fission source is precomputed before converging the scattering source.