File path bug fix in WritePVTU
When writing the PVTU file during the exportation of a field function,
each Piece Source included any full or relative path included in the
base_filename variable. Since the pvtu and vtu files reside in the same
folder, this path is incorrect and should be omitted to allow paraview
to locate the vtu files when reading the pvtu. This commit corrects this
issue.